+++ forte2/ci/__init__.py
@@ -1,4 +1,5 @@
-from .ci import CISolver, CI, RelCISolver, RelCI
+from .ci import CISolver, CI
+from .rel_ci import RelCISolver, RelCI
 from .ci_utils import (
     pretty_print_ci_summary,
     pretty_print_ci_nat_occ_numbers,
