+++ forte2/dsrg/rel_dsrg_mrpt2_slow.py
@@ -63,7 +63,7 @@ class RelDSRG_MRPT2_Slow(DSRGBase):
 
     def __post_init__(self):
         super().__post_init__()
-        self.requires_attrs |= {"two_component": True}
+        self.requires_attrs.update({"two_component": True})
 
     def get_integrals(self):
         g1, g2, l2, l3 = self.ci_solver.make_average_cumulants()
