+++ forte2/sci/__init__.py
@@ -1,6 +1,2 @@
-from .sci import (
-    SelectedCI,
-    SelectedCISolver,
-    RelSelectedCI,
-    RelSelectedCISolver,
-)
+from .sci import SelectedCI, SelectedCISolver
+from .rel_sci import RelSelectedCI, RelSelectedCISolver
