+++ tests/ci/test_rel_ci_rdms.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from forte2 import State, MOSpace
 from forte2.jkbuilder import SpinorbitalIntegrals
-from forte2.ci.ci import _RelCISingleStateSolver
+from forte2.ci.rel_ci import _RelCISingleStateSolver
 from forte2.helpers.comparisons import approx
 
 rdm_threshold = 1e-12
