+++ tests/mcopt/test_sa_casscf_n2.py
@@ -69,15 +69,19 @@ def test_sa_mcscf_diff_mult_with_avas():
 
     nat_occ_ref = np.array(
         [
-            [1.97531261, 1.9761829, 1.9797172, 1.97636293],
-            [1.91200523, 1.4525388, 1.46608628, 1.5689295],
-            [1.91200523, 1.45253877, 1.46608627, 1.56892948],
-            [0.08779586, 0.54735532, 0.53380304, 0.43094082],
-            [0.08779586, 0.54735529, 0.53380303, 0.4309408],
-            [0.0250852, 0.02402894, 0.02050418, 0.02389647],
+            [1.97531261, 1.9761829, 1.9797172],
+            [1.91200523, 1.4525388, 1.46608628],
+            [1.91200523, 1.45253877, 1.46608627],
+            [0.08779586, 0.54735532, 0.53380304],
+            [0.08779586, 0.54735529, 0.53380303],
+            [0.0250852, 0.02402894, 0.02050418],
         ]
     )
+    nat_occ_avg_ref = np.array(
+        [1.97636293, 1.5689295, 1.56892948, 0.43094082, 0.4309408, 0.02389647]
+    )
     assert mc.ci_solver.nat_occs == pytest.approx(nat_occ_ref, abs=5e-7)
+    assert mc.ci_solver.nat_occs_avg == pytest.approx(nat_occ_avg_ref, abs=5e-7)
 
 
 def test_sa_casscf_diff_mult():
